+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Changed
     - Move ``setup_ip_allowing()`` to ``IPAllowMixin`` (#1237)
     - Improve ``syncmodifyapi`` project logging (#1228)
     - Do not exit ``syncmodifyapi`` on failure (#1229)
+    - Simplify ``syncmodifyapi`` project querying (#1264)
 - **Timeline**
     - Handle app plugin exceptions in ``get_object_link()`` (#1232)